requirements.txt
python-dotenv
chainlit==0.5.2
langchain
huggingface_hub
